image-hd: do not use first partition offset for GPT's first usable LBA
Currently first usable LBA in the GPT header is pointing to offset of
the first partition, ignoring the gpt-location specified in the config.
This can lead to some issues as explained in [1]. Disabling this
behavior doesn't break any tests and allows for generating of images
that have same layout as those generated by sgdisk or other utilities.

[1] https://github.com/pengutronix/genimage/issues/262

Signed-off-by: Jan Čermák <sairon@sairon.cz>