image-hd: set the first usable lba correctly
Before this, the begining of the first partition was used. This leaves
some unusable space between the partition table and the first partition
and causes some Windows versions to alter the GPT.

So set the first usable lba correctly: Set it to the first lba after the
last non-partition data before the first partition.
This will either be directly behind the GPT entry array or, for example
after a bootloader that is placed between the partition table and the
first partition.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>