image-hd: add sparse option for partitions
When enabled (the default and previous state), any holes in the input
image will be copied to the hdimage and the space after the partition
image and before the next partition is also created as a hole.

When disabled, all holes within the partition image are filled with
zeros while the image is copied. If the 'fill' option is set as well,
then the space after the partition image and before the next partition
is also filled with zeros.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>