image-fip: check size of generated image
Until now, the image->size of the generated FIP image was 0, which meant
that hdimage_generate() did not check if the image fits into the
available space of a partition, which could result in an incomplete FIP
image if the image was too large and would be overwritten by contents of
later partitions. Determine the size of the generated FIP image with a
stat() call, so that hdimage_generate() can error out if it does not
fit, e.g. in our test case:

  INFO: hdimage(test.hdimage): adding primary partition 'part1' (in MBR) from 'test.fip' ...
  ERROR: hdimage(test.hdimage): part part1 size (512) too small for test.fip (12992)

Signed-off-by: Roland Hieber <rhi@pengutronix.de>