config: add new randomseed option
Add a new 'randomseed' option that allows for reproducible UUIDs
without needing to manually specify them, which is helpful for
implementing reproducible builds and delta updates.

When randomseed is supplied, the random-number generator is switched
from a libc's random() to a user-seeded PRNG that uses the Marsaglia
xorwow algorithm.

If it is not supplied, genimage continues uses random() with a
timestamp seed.

To use this feature, set the randomseed config option to any arbitrary
string (truncated at 4096 characters). If left unset, the PRNG will be
seeded by the current time's microseconds field as before.

Signed-off-by: Nicholas Clark <nicholas.clark@gmail.com>