disk images: fix some corner-cases in MBR calculation
Fix some corner-cases with CHS calculation when filling out MBR
partition tables. Make sure that CHS values are correctly clamped
the requested LBA is too big to be representable as a CHS, and fix
the calculation when LBA=0.

When creating a hybrid GPT/MBR, change the GPT protective partition
size from (hd->gpt_location / 512 + GPT_SECTORS - 2) to GPT_SECTORS,
which is more generally correct (and evaluates to the same value
in the hybrid case).

Signed-off-by: Nicholas Clark <nicholas.clark@gmail.com>