genimage: allow non-sparse image-less partitions not in partition table
Previously, partitions not included in a partition table were required
to always reference an image.

However, it can be useful to define such partitions without an image,
for example, for "barebox-env" or "barebox-state" regions at fixed
offsets in legacy MBR setups. These partitions are typically populated at
runtime and therefore do not need an image.

To ensure that flashing a sparse image variant results in a well-defined
state, these partitions should be explicitly filled with zeros to
overwrite any existing data.

So allow partitions not present in the partition table, without an
image, to be filled with zeros, e.g.:

  partition barebox-environment {
    size = 128K
    in-partition-table = false
    sparse = false
  }

Signed-off-by: Bastian Krause <bst@pengutronix.de>