test: cpio: add variant that should fail due to missing cpio util
Genimage will currently exit with a success exit code when generating a
cpio archive fails due to a missing `cpio` util.
This is due to a pipe in the `systemp` call and the `pipefail` option not
being set.

See this issue[1] for mor information.

Add a test that checks if genimage fails if the cpio util is missing
(which it currently does not but should).

[1]: https://github.com/pengutronix/genimage/issues/320

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>